requirements.txt
albumentations==2.0.8
clip==0.2.0
comet_ml==3.49.11
coremltools==8.3.0
dvclive==3.48.2
flash_attn==2.8.0.post2
flatbuffers==24.12.23
gradio==5.34.2
hub_sdk==0.0.20
huggingface_hub==0.27.0
ipython==8.12.3
lap==0.5.12
lvis==0.5.3
matplotlib==3.5.3
mct_quantizers==1.6.0
MNN==3.2.1
model_compression_toolkit==2.3.0
mss==10.0.0
ncnn==1.0.20250503
nncf==2.17.0
numpy==1.24.4
onnx==1.17.0
onnx2tf==1.28.0
onnxruntime==1.19.2
onnxslim==0.1.57
opencv_contrib_python==4.11.0.86
openvino==2025.2.0
paddle==1.0.2
pafy==0.5.5
pandas==2.0.3
Pillow==11.2.1
pillow_heif==0.22.0
psutil==6.0.0
py_cpuinfo==9.0.0
pycocotools==2.0.10
pytest==8.4.1
pytubefix==9.2.0
PyYAML==6.0.2
PyYAML==6.0.2
ray==2.47.1
Requests==2.32.4
roboflow==1.1.66
sahi==0.11.23
scipy==1.15.3
seaborn==0.13.2
sentry_sdk==2.17.0
Shapely==2.1.1
sony_custom_layers==0.3.0
streamlit==1.46.0
super_gradients==3.7.1
tensorflow==2.19.0
tensorflowjs==4.22.0
tensorrt==10.12.0.36
tflite_runtime==2.14.0
tflite_support==0.4.4
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.0
transformers==4.52.4
tritonclient==2.58.0
ultralytics_thop==2.0.13
wandb==0.20.1
x2paddle==1.6.0
yt_dlp==2025.6.9
